Update ListView.md: iconFieldName description
I had a hard time understanding how to use iconFieldName and its effect. It was not clear to me by the original description that it required the property name of the items with the path of the file and that it would automagically create the entire column, header and field rendering. I had to read the source code to figure it out.