Support for PDF icons
I realize that it may not be the cleanest approach and would not be shocked if it is not accepted, but here is a workaround that I propose to deal with PDF icons which don't exist in the "image" flavour of FileTypeIcon. I find it annoying that PDF files have the empty icons.
Ideally, I could have access to FileTypeIcon's private _getFileExtension method for a more robust assessment of the PDF extension.