Update ChartControl.tsx
The _initChart function is not initialising the chart options properly. The options passed as a parameter are never used in the function which means that, when the function is being called from ChartControl.componentWillReceiveProps, the control is never updated with the new properties but it will use the old properties instead.