fix(taxonomy tree): options remain selected when switching in single mode
This is a hack fix since the tree is already built on top of React anti-patterns, which cause this in the first place. In the future a full refactor is really required to sort this out at its roots.

By adding an unstable `key`, it forces the component to re-render on every render. Without this, React assumes nothing has changed and does not react to the `selectedKey` update.

For large trees with everything expanded, this could cause performance issues but should be okay in most cases.